User model gets pfp
Fixed an issue where UserRole and ProfilePictureBase64 were not properly binding in the UserController

Fixed an issue where the "Account" link did not always go to the user account page when the user was an instructor

Created new column for "UserProfilePictureBase64" in the sql server; Integrated into model and migrated.

Previous users now have a default image for their table value.

Users now receive a default image for their profile upon signing up

new method GetProfileImage translate image from Base64 to img/png and sets it to the "account" tab

New action method GET ChangeProfilePicture for calling the view

New action method POST for uploading and posting a new pfp to the DB as a Base64 String