Bug Fixes; Totally awesome dude.
New UI for the instructor grading page.

Fixed an outdated method for editing a user account.

Fixed a bug that caused EF Tracking in the authorization interface implementation

Fixed an issue that would cause the UpdateAdd method to throw an invalid model error

Rebuilt an outdated pfp updating method. Fixed a bug that caused the pfp post method to never be called.
(STILL ONE PROBLEM!!!)